+++ genomicsurveillance/models/__init__.py
@@ -1,13 +1,5 @@
-from .models import (
-    MultiLineage,
-    MultiLineageArma,
-    MultiLineageClockReset,
-    SimpleMultiLineage,
-)
+from .clock_reset import MultiLineageClockReset as MultiLineage
 
 __all__ = [
     "MultiLineage",
-    "SimpleMultiLineage",
-    "MultiLineageArma",
-    "MultiLineageClockReset",
 ]
